Fix reading memberOf from the user ID Overrride for non-admins
memberOf attribute was not on the list of attributes that could be
read out of user ID Override by any authenticated user other than
a member of admins group.

This commit adds 'memberof' to the corresponding permission so that
all authenticated users will be able to read memberOf attribute now.
This allows Web UI to see memberOf of an Active Directory user to
decide whether self-service page should or should not be shown even
if that user is not a member of 'admins' group.